checked install script so no longer requires git

+++ install.sh
@@ -10,17 +10,19 @@ if [ -z "$1" ]; then
     exit 1
 fi
 
-git clone https://github.com/screencom/systemd-ngrok.git
-cd systemd-ngrok
-cp ngrok.service /lib/systemd/system/
 mkdir -p /opt/ngrok
-cp ngrok.yml /opt/ngrok
-sed -i "s/<add_your_token_here>/$1/g" /opt/ngrok/ngrok.yml
 
 cd /opt/ngrok
 wget https://bin.equinox.io/c/4VmDzA7iaHb/ngrok-stable-linux-arm.zip
-unzip ngrok-stable-linux-arm.zip
-rm -f ngrok-stable-linux-arm.zip
+
+if [[ ! -f ngrok-stable-linux-arm.zip ]]; then
+    unzip ngrok-stable-linux-arm.zip
+    rm -f ngrok-stable-linux-arm.zip
+fi
+
+curl -so /lib/systemd/system/ngrok.service https://github.com/screencom/systemd-ngrok/raw/master/ngrok.service
+curl -so /opt/ngrok/ngrok.yml https://github.com/screencom/systemd-ngrok/raw/master/ngrok.yml
+sed -i "s/<add_your_token_here>/$1/g" /opt/ngrok/ngrok.yml
 
 systemctl enable ngrok.service
 systemctl start ngrok.service
